+++ server/controller/src/test/java/ai/starwhale/mlops/domain/runtime/RuntimeServiceTest.java
@@ -70,7 +70,6 @@ import ai.starwhale.mlops.domain.runtime.po.RuntimeEntity;
 import ai.starwhale.mlops.domain.runtime.po.RuntimeVersionEntity;
 import ai.starwhale.mlops.domain.storage.StoragePathCoordinator;
 import ai.starwhale.mlops.domain.storage.StorageService;
-import ai.starwhale.mlops.domain.system.SystemSetting;
 import ai.starwhale.mlops.domain.trash.TrashService;
 import ai.starwhale.mlops.domain.user.UserService;
 import ai.starwhale.mlops.domain.user.bo.User;
@@ -91,7 +90,6 @@ import io.kubernetes.client.openapi.models.V1JobSpec;
 import io.kubernetes.client.openapi.models.V1ObjectMeta;
 import io.kubernetes.client.openapi.models.V1PodSpec;
 import io.kubernetes.client.openapi.models.V1PodTemplateSpec;
-import io.kubernetes.client.openapi.models.V1Secret;
 import java.io.IOException;
 import java.util.List;
 import java.util.Map;
@@ -557,32 +555,6 @@ public class RuntimeServiceTest {
 
         res = service.query("p1", "r1", "v3");
         assertThat(res, is("n3"));
-    }
-
-    @Test
-    public void testUpdateDockerSettingWithNone() throws ApiException {
-        service.onUpdate(new SystemSetting());
-        verify(k8sClient, times(1)).deleteSecret(anyString());
-    }
-
-    @Test
-    public void testUpdateDockerSettingWithValueAndExistedSecret() throws ApiException {
-        given(k8sClient.getSecret(anyString())).willReturn(new V1Secret());
-        var setting = new SystemSetting();
-        setting.setDockerSetting(new DockerSetting("mockRegistry", "admin", "admin"));
-        service.onUpdate(setting);
-
-        verify(k8sClient, times(1)).replaceSecret(anyString(), any());
-    }
-
-    @Test
-    public void testUpdateDockerSettingWithValueAndNonExistedSecret() throws ApiException {
-        given(k8sClient.getSecret(anyString())).willThrow(new ApiException(HttpServletResponse.SC_NOT_FOUND, ""));
-        var setting = new SystemSetting();
-        setting.setDockerSetting(new DockerSetting("mockRegistry", "admin", "admin"));
-        service.onUpdate(setting);
-
-        verify(k8sClient, times(1)).createSecret(any());
     }
 
     @Test
